refactor: rename helpers file to common.ts
