+++ src/app/utils/field-validation/validators/sectionValidator.ts
@@ -3,7 +3,7 @@ import { Either, left, right } from 'fp-ts/lib/Either'
 import { ProcessedFieldResponse } from 'src/app/modules/submission/submission.types'
 import { ResponseValidator } from 'src/types/field/utils/validation'
 
-import { isAnswerEmpty } from './helpers'
+import { isAnswerEmpty } from './common'
 
 /**
  * A function that returns a validation function for a section field when called.
